refactor: modularize ethereum staking page and related functionality
