+++ frontend/app/src/modules/staking/eth/components/EthStakingPagePlaceholder.vue
