+++ frontend/app/src/modules/staking/eth/components/EthValidatorCombinedFilter.vue
