+++ frontend/app/src/modules/staking/eth/components/EthValidatorFilter.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type { EthStakingCombinedFilter, EthStakingFilter, EthStakingFilterType } from '@rotki/common';
 import Eth2ValidatorFilter from '@/components/helper/filter/Eth2ValidatorFilter.vue';
-import EthValidatorCombinedFilter from '@/components/staking/eth/EthValidatorCombinedFilter.vue';
+import EthValidatorCombinedFilter from './EthValidatorCombinedFilter.vue';
 
 const filterModel = defineModel<EthStakingCombinedFilter | undefined>('filter', { required: true });
 const model = defineModel<EthStakingFilter>({ required: true });
