+++ frontend/app/src/pages/staking/[[location]].vue
@@ -29,7 +29,7 @@ const props = defineProps<{
 const imageSize = '64px';
 
 const pages = {
-  eth2: defineAsyncComponent(() => import('@/components/staking/eth/EthStakingPage.vue')),
+  eth2: defineAsyncComponent(() => import('@/modules/staking/eth/EthStakingPage.vue')),
   kraken: defineAsyncComponent(() => import('@/components/staking/kraken/KrakenPage.vue')),
   liquity: defineAsyncComponent(() => import('@/components/staking/liquity/LiquityPage.vue')),
 };
